back/requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2023.11.17
click==8.1.7
confluent-kafka==2.3.0
contextlib2==21.6.0
ecdsa==0.18.0
fastapi==0.109.0
fastapi-cache2==0.2.1
fastapi-filter==1.1.0
fastapi-pagination==0.12.14
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
jsonpickle==3.0.2
kafka-python==2.0.2
loguru==0.7.2
multidict==6.0.4
opencv-contrib-python==4.9.0.80
packaging==23.2
passlib==1.7.4
pendulum==3.0.0
pillow==10.2.0
pluggy==1.4.0
priority==2.0.0
pyasn1==0.5.1
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-schema==0.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
schema==0.7.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.27.0.post1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
